py-services/requirements.txt
# Web框架和异步支持
starlette>=0.27.0,<0.36.0
httpx>=0.24.0,<0.27.0
fastapi>=0.100.0,<0.110.0
uvicorn>=0.24.0,<0.30.0
pydantic>=2.0.0,<3.0.0

# 任务队列
celery>=5.3.0,<5.4.0
redis<6.0.0,>=5.0.1
async_timeout>=4.0.0

# LlamaIndex核心和扩展
llama-index==0.10.68
llama-index-readers-file==0.1.33
llama-index-embeddings-huggingface==0.1.5
llama-index-vector-stores-redis==0.1.2

# 文档解析
pypdf>=4.0.0,<5.0.0
markdown>=3.5.0,<4.0.0
python-docx>=1.0.0,<2.0.0
python-pptx>=0.6.21,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
pymupdf>=1.23.0,<2.0.0

# 图像处理(用于OCR)
pdf2image>=1.16.0,<2.0.0
pytesseract>=0.3.10,<0.4.0

# 向量嵌入
transformers>=4.30.0,<5.0.0
sentence-transformers>=2.2.0,<3.0.0
torch>=2.1.0,<3.0.0; python_version >= "3.11"

# 数据处理
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0

# 工具库
python-dotenv>=1.0.0,<2.0.0
loguru>=0.7.0,<1.0.0
requests>=2.31.0,<3.0.0
tenacity>=8.2.0,<9.0.0  # 用于重试机制
python-multipart>=0.0.5,<1.0.0

# 测试工具
pytest>=7.3.1,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0

# 大模型和向量API
dashscope>=1.10.0,<2.0.0       # 通义千问API调用
openai>=1.6.0,<2.0.0           # OpenAI兼容接口

# 文本处理
nltk>=3.8.1,<4.0.0             # 自然语言处理工具库，用于句子分割
chardet>=5.1.0,<6.0.0          # 文本编码检测
langdetect>=1.0.9,<2.0.0       # 语言检测
jieba>=0.42.1,<0.43.0          # 中文分词
fugashi>=1.2.0,<2.0.0          # 日语文本处理

# 系统监控和诊断
psutil>=5.9.0,<6.0.0           # 系统信息获取

# 文件存储
minio>=7.1.15,<8.0.0           # MinIO客户端